Track segment coloring is easily user configurable - per active session (sometimes optimal range is 4-7, sometimes 7-20 minutes per km).

+++ app/src/main/java/ee/taltech/likutt/iti0213_2019s_hw02/LocationService.kt
@@ -64,8 +64,8 @@ class LocationService : Service() {
 
     private val dateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSXXX", Locale.getDefault())
 
-    private var minSpeed: Int = 6*60  // 1 km within 6 min  -- minspeed=6*60 km / per sec
-    private var maxSpeed: Int = 18*60
+    private var minSpeed: Double = (6*60).toDouble()  // 1 km within 6 min  -- minspeed=6*60 km / per sec
+    private var maxSpeed: Double = (18*60).toDouble()
 
     private var prevLocation: Location? = null
     private var prevTime: Long? = null
